Move Raytracing from Tracker to ExplorationTracker
Makes patrolling so much faster
