+++ Assets/Scripts/Trackers/PatrollingTracker.cs
@@ -80,7 +80,7 @@ namespace Maes.Trackers
             }
         }
 
-        protected override void OnLogicUpdate(IReadOnlyList<MonaRobot> robots)
+        protected override void OnLogicUpdate(MonaRobot[] robots)
         {
             var worstGraphIdleness = 0;
             var graphIdlenessSum = 0;
